requirements.txt
filelock==3.16.1
fsspec==2024.10.0
jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
pillow==10.4.0
ruff==0.7.1
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
triton==3.0.0
typing-extensions==4.12.2
